subnetcalculator/requirements.txt
pydantic>=2.7.2,<3
PyYaml>=6.0.0,<7.0.0